requirements.txt

streamlit==1.28.1
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
opencv-python==4.8.0.76
tqdm==4.65.0
kaggle==1.5.16
plotly==5.15.0
